Preparing version Alpha 1.0

+++ Linked/Game.cpp
@@ -71,7 +71,7 @@ Game::Game(int windowsWidth, int windowsHeight)
 				//{
 				if (!MapTerrainImageLoader::isOfCollisionType(monsterMap->getMapCoordinateForPlayerMovement(glm::vec3(i, j, 0)).terrain))
 				{
-					monster->getTransform()->translate(i, j, 1.0f);
+					monster->getTransform()->translate((float)i, (float)j, 1.0f);
 					monsters.push_back(monster);
 				}
 				//aux = true;
